refactored open status logic into util and added testing
